Move account methods to separate files
