+++ src/types/Account/methods/getIdentityIds.js
@@ -1,0 +1,9 @@
+/**
+ *
+ * @return {string[]}
+ */
+function getIdentityIds() {
+  return this.storage.getIndexedIdentityIds(this.walletId).filter(Boolean);
+}
+
+module.exports = getIdentityIds;
